Updating GitHub links:
The build file will now attempt to get the latest semantic tag first, on failure it will attempt to get the commit hash, and on failure it will use Not Found as ldflags to the Ginkgo build.
common.go now contains the code which will attempt to set the commit hash, or Not Found.
ptp_suite_test.go now checks that the origin_url has been set by the ldflags, otherwise it calls the function in common.go to set the commit hash. This is useful if running the Ginkgo test directly.
Not Found is, in my view, preferable to a link to the main branch as that may be ambiguous, but opinions on this are welcome.